handle race condition better, in case file gets deleted after we check that it exists

+++ sarracenia/flow/__init__.py
@@ -2021,12 +2021,12 @@ class Flow:
 
             # assert new_inflight_path is set.
 
-            if os.path.exists(msg['new_inflight_path']):
+            if self.o.inflight and os.path.exists(msg['new_inflight_path']):
 
-                if self.o.inflight:
+                try:
                     how_old = time.time() - os.path.getmtime(msg['new_inflight_path'])
-                    #FIXME: if mtime > 5 minutes, perhaps rm it, and continue? what if transfer crashed?
-                    #       Added this with fixed value, should it be a setting?
+                    # if file exists and mtime > 5 minutes, rm it, and continue. previous transfer likely crashed
+                    #FIXME: Added this with fixed value, should it be a setting?
                     if how_old > 300:
                         os.unlink( msg['new_inflight_path'] )
                         logger.info(
@@ -2034,8 +2034,13 @@ class Flow:
                     else:
                         logger.warning(
                             f"inflight file already exists. race condition, deferring transfer of {msg['new_path']}" )
-                    self.worklist.failed.append(msg)
-                    continue
+                # in case file gets deleted after the exists check but before the getmtime or unlink calls
+                except FileNotFoundError:
+                    logger.warning(
+                        f"inflight file already existed. race condition, deferring transfer of {msg['new_path']}")
+
+                self.worklist.failed.append(msg)
+                continue
                 # overwriting existing file.
 
             # FIXME: decision of whether to download, goes here.
